refactor for separate hooks
